Rename to Consumer and Memento.
Rename the function `async` to `Consumer` to create a consumer object
to use to asynchronously iterate the queue. Rename the function `sync`
to `memento` to create a marker to a place in the queue that can be used
to create a `Consumer` that begins at a specific spot in the queue.