Externalize windows and throttles.
Migrating Procession to become merely a linked list, removing the
concept of listeners from Procession itself. The Window class will
maintain a trailing shifter and the Throttle class will implement an
enqueue that will push back if a limit is reached.

This allows for subsequent simplification of Procession when downstream
libraries no longer depend on it's ability to buffer messages itself and
use Window and Throttle to create reservoirs and backlogs.