Reader/Writer instead of "Socket."
Created a Reader/Writer pair for asynchronous serialization. They are
simple and based on the new Staccato. Upgraded Stacacto to 8.0.x.

Error handling is probably not completely covered by the Writer. I've
found that I often top writing before an underlying socket is done
complaining leading to unhandled `EventEmitter` errors. The new strategy
is to stop processing and treat that as a truncate, and have error
handling be a separate affair that probably also stops the Writer.